// app.client/src/api/models/CustomerQuoteListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerQuoteListVM = {
    customerQuoteId?: string | null;
    quoteToOrderPo?: string | null;
    quoteName?: string | null;
    customerId?: string | null;
    customerShippingAddressId?: string | null;
    warehouseId?: string | null;
    freightTermId?: string | null;
    shipMethodId?: string | null;
    shipToAttention?: string | null;
    shipToAccountId?: string | null;
    quoteToContactId?: string | null;
    shippingAccountNum?: string | null;
    quoteDate?: string;
    doesQuoteExpire?: boolean;
    expiresOn?: string;
    isExpired?: boolean;
    headerComment?: string | null;
    quoteToContact?: string | null;
    quoteToEmail?: string | null;
    quoteToCellProvider?: string | null;
    quoteToCellNumber?: string | null;
    quotedByEmail?: string | null;
    quotedByFullName?: string | null;
    trackingNumEmail?: string | null;
    trackingNumFullName?: string | null;
    displayForCustomer?: boolean;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};